agentic_Rag/requirements.txt
streamlit
faiss-cpu
sentence-transformers
python-docx
pdfplumber
python-pptx
pandas
markdown
tqdm